requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langgraph>=0.0.30

chromadb>=0.4.0
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0

pdf2image>=1.16.0
easyocr>=1.7.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

openai-whisper>=20231117

python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0